--- packages/server/src/models/index.ts
@@ -1,5 +1,0 @@
-export * from './posts.model';
-export * from './site-theme.model';
-export * from './threads.model';
-export * from './user.model';
-export * from './user-theme.model';
