+++ packages/server/src/post/dto/index.ts
@@ -1,0 +1,1 @@
+export {}
