+++ packages/server/src/post/index.ts
@@ -1,0 +1,2 @@
+export * from './dto';
+export * from './post.model';
