+++ packages/server/src/site-theme/dto/index.ts
@@ -1,0 +1,1 @@
+export {}
