+++ packages/server/src/thread/index.ts
@@ -1,0 +1,2 @@
+export * from './dto';
+export * from './thread.model';
