+++ packages/server/src/user-theme/dto/index.ts
@@ -1,0 +1,1 @@
+export {}
