+++ packages/server/src/user-theme/index.ts
@@ -1,0 +1,2 @@
+export * from './dto';
+export * from './user-theme.model';
