+++ packages/server/src/user/dto/index.ts
@@ -1,0 +1,1 @@
+export {}
